Taking the project another direction
