--- config.yaml
@@ -1,2 +1,0 @@
-logging:
-  level: DEBUG
