+++ model_inspect/__main__.py
@@ -1,7 +1,7 @@
-from model_inspect.cli import cli
+"""
+Main Entrypoint into the module
+"""
+from model_inspect.cli import main
 
 if __name__ == "__main__":
-    # standalone_mode=False:
-    # This tells Click not to handle exceptions or exit on its own,
-    # giving you control over the output.
-    cli(prog_name="modelinspect", standalone_mode=False)
+    main()
