--- model_inspect/file_types/base_file/__init__.py
@@ -1,14 +1,0 @@
-"""
-* File Path: Path to the binary file.
-* Start Address: The starting address of the file in bytes.
-* End Address: The ending address of the file in bytes.
-* File Size: The total size of the binary file.
-* Magic Number: A unique identifier used to recognize file types.
-* File Type: A descriptor or identifier for the file type (e.g., ZIP, ELF).
-* Checksum/Hash: To verify file integrity (e.g., MD5, SHA-256).
-* Permissions: Any file-level permissions or security settings.
-* Timestamp: The creation or modification timestamp.
-* Header/Footer Offsets: Positions in the file where headers and footers reside.
-* Data Sections: Markers for different sections if the file format includes logical partitions.
-* Endianness: The byte order for interpreting multibyte numbers.
-"""
