--- model_inspect/file_types/zip_file/__init__.py
@@ -1,28 +1,0 @@
-""" File Type: Zip File Init """
-
-from model_inspect.file_types.zip_file.constants import (
-    ZIP_BZIP2,
-    ZIP_DEFLATED,
-    ZIP_LZMA,
-    ZIP_STORED,
-)
-
-# from .compression import Compressor
-# from .decompression import Decompressor
-# from .encryption import ZipDecrypter
-# from .exceptions import ZipFileError
-# from .zip_info import ZipInfo
-# from .zip_reader import ZipExtFile
-
-__all__ = [
-    "ZipInfo",
-    "ZipExtFile",
-    "Compressor",
-    "Decompressor",
-    "ZipDecrypter",
-    "ZipFileError",
-    "ZIP_STORED",
-    "ZIP_DEFLATED",
-    "ZIP_BZIP2",
-    "ZIP_LZMA",
-]
