--- model_inspect/file_types/zip_file/exceptions.py
@@ -1,2 +1,0 @@
-class BadZipFile(Exception):
-    """Raised when a ZIP file is found to be corrupt."""
